backend/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
python-dotenv==1.2.1
sqlalchemy==2.0.46
psycopg2-binary==2.9.9
passlib==1.7.4
python-jose==3.5.0
email-validator==2.3.0
python-multipart==0.0.22
bcrypt==3.2.0
celery==5.3.6
redis==5.0.1
httpx==0.27.0
reportlab==4.2.5
